+++ src/include/access/htup_details.h
@@ -482,7 +482,7 @@ HeapTupleHeaderSetMovedPartitions(HeapTupleHeaderData *tup)
 static inline uint32
 HeapTupleHeaderGetDatumLength(const HeapTupleHeaderData *tup)
 {
-	return VARSIZE(tup);
+	return (uint32) VARSIZE(tup);
 }
 
 static inline void
